refactor: resolve PR #307 review comments
