+++ platforms/godot_editor/addons/admob/csharp/src/mock/MockAppOpenAdPlugin.cs.uid
@@ -1,0 +1,1 @@
+uid://bvi5ufcupyusr
