--- platforms/godot_editor/addons/admob/csharp/src/mock/MockFullScreenAdPlugin.cs.uid
@@ -1,1 +1,0 @@
-uid://km6rc08mggh5
