+++ platforms/godot_editor/addons/admob/csharp/src/mock/MockInterstitialAdPlugin.cs.uid
@@ -1,0 +1,1 @@
+uid://0ugv4fypxu3o
