+++ platforms/godot_editor/addons/admob/internal/mock/mock_ad_size_plugin.gd
@@ -1,17 +1,40 @@
+# MIT License
+#
+# Copyright (c) 2023-present Poing Studios
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+# gdlint: disable=function-name
 extends Node
 
-# gdlint: disable=function-name
+
 func getSmartBannerAdSize() -> Dictionary:
 	return {"width": 320, "height": 50}
-# gdlint: disable=function-name
+
 
 func getCurrentOrientationAnchoredAdaptiveBannerAdSize(width: int) -> Dictionary:
-# gdlint: disable=function-name
 	return {"width": width, "height": 50}
 
-# gdlint: disable=function-name
+
 func getPortraitAnchoredAdaptiveBannerAdSize(width: int) -> Dictionary:
 	return {"width": width, "height": 50}
+
 
 func getLandscapeAnchoredAdaptiveBannerAdSize(width: int) -> Dictionary:
 	return {"width": width, "height": 50}
