+++ platforms/godot_editor/addons/admob/internal/mock/mock_app_open_ad_plugin.gd.uid
@@ -1,0 +1,1 @@
+uid://tnr2tyn44yii
