--- platforms/godot_editor/addons/admob/internal/mock/mock_full_screen_ad_plugin.gd.uid
@@ -1,1 +1,0 @@
-uid://bp5la8gkubu1g
