+++ platforms/godot_editor/addons/admob/internal/mock/mock_rewarded_interstitial_ad_plugin.gd.uid
@@ -1,0 +1,1 @@
+uid://6yu2agtfkqgo
